Creando nuevo constructor de consultas
